+++ src/static/code-editor.js
@@ -39,3 +39,30 @@ function json2css(json)
 	// Return beautiful css :)
 	return css.trim();
 }
+
+
+// Function to check if LESS variables are present in <mbr-parameters>
+function checkLessVars(html, css) {
+	// Extract variables from LESS
+	let regex = /@[a-zA-Z0-9_\-{]+/gm;
+	let vars = [];
+	vars = css.match(regex);
+
+	// Clean up found vars
+	vars.forEach((v, i) => {
+		vars[i] = v.replace(/@/g, "").replace(/{/g, "").replace(/^bg-.+$/, "bg");
+	});
+
+	// Remove duplicates
+	vars = vars.filter((a, b) => vars.indexOf(a) === b)
+
+	// Check if found LESS vars exist in <mbr-parameters>
+	let noexist = [];
+	html = $(html);
+	vars.forEach(function(v) {
+	  if ( html.find("mbr-parameters [name='" + v + "']").length == 0 )
+		noexist[noexist.length] = v;
+	});
+
+	return (noexist.length === 0 ? false : noexist);
+}